Reorganised code; removed dead code; renamed byte code dumper
