+++ functionid/src/main/java/tech/pegasys/poc/witnesscodeanalysis/datafile/ContractData.java
@@ -1,4 +1,4 @@
-package tech.pegasys.poc.witnesscodeanalysis;
+package tech.pegasys.poc.witnesscodeanalysis.datafile;
 
 import java.util.List;
